+++ src/server/metiers/configuration/dependencies.container.ts
@@ -1,4 +1,4 @@
-import { MétierRepository } from '~/server/metiers/domain/métier.repository';
+import { MétierRepository } from '~/server/metiers/domain/metierAlternance.repository';
 import { RécupérerMétiersUseCase } from '~/server/metiers/useCases/récupererMétiersUseCase';
 
 export interface MétierDependencies {
