+++ src/server/metiers/domain/metierAlternance.repository.ts
@@ -1,6 +1,6 @@
 import { Either } from '~/server/errors/either';
 
-import { MetierAlternance } from './metier';
+import { MetierAlternance } from './metierAlternance';
 
 export interface MétierRepository {
 	getMetierList(recherche: string): Promise<Either<Array<MetierAlternance>>>
