+++ src/server/metiers/infra/apiLaBonneAlternanceMétier.mapper.ts
@@ -1,5 +1,5 @@
 import { MetierLaBonneAlternanceApiResponse } from '~/server/alternances/infra/repositories/apiLaBonneAlternance';
-import { MetierAlternance } from '~/server/metiers/domain/metier';
+import { MetierAlternance } from '~/server/metiers/domain/metierAlternance';
 
 export const mapMetier = (response: MetierLaBonneAlternanceApiResponse): Array<MetierAlternance> => {
 	const listeMetiers = response.labelsAndRomes;
