+++ src/server/metiers/infra/apiLaBonneAlternanceMétier.repository.ts
@@ -1,7 +1,7 @@
 import { MetierLaBonneAlternanceApiResponse } from '~/server/alternances/infra/repositories/apiLaBonneAlternance';
 import { createSuccess, Either } from '~/server/errors/either';
-import { MetierAlternance } from '~/server/metiers/domain/metier';
-import { MétierRepository } from '~/server/metiers/domain/métier.repository';
+import { MetierAlternance } from '~/server/metiers/domain/metierAlternance';
+import { MétierRepository } from '~/server/metiers/domain/metierAlternance.repository';
 import { mapMetier } from '~/server/metiers/infra/apiLaBonneAlternanceMétier.mapper';
 import { ErrorManagementService } from '~/server/services/error/errorManagement.service';
 import { PublicHttpClientService } from '~/server/services/http/publicHttpClient.service';
